WIP Refactored.
