+++ .envrc
@@ -1,0 +1,7 @@
+# For java
+export MONGO_HOST="localhost:27017"
+export MONGO_DB="literature"
+export MONGO_COLLECTION="quotes"
+export MONGO_USER="root"
+export MONGO_PASSWORD="root"
+export PIPELINE_CLASS="net.yuwtennis.app.pipelines.MongoIOSimpleRWQueryService"
