--- env.sh
@@ -1,3 +1,0 @@
-export MONGO_URI="mongodb://localhost:27017"
-export MONGO_DB="literature"
-export MONGO_COLLECTION="quotes"
