+++ java/Makefile
@@ -1,0 +1,9 @@
+
+clean:
+	mvn clean
+
+package:
+	mvn package
+
+run:
+	java -cp target/pipeline_apps-1.0-SNAPSHOT-jar-with-dependencies.jar net.yuwtennis.app.AppClient
