+++ java/src/main/java/net/yuwtennis/app/repositories/MongoIO.java
@@ -1,4 +1,4 @@
-package net.yuwtennis.app.helpers.ios;
+package net.yuwtennis.app.repositories;
 
 import org.apache.beam.sdk.io.mongodb.MongoDbIO;
 import org.apache.beam.sdk.values.PCollection;
