+++ java/src/test/java/net/yuwtennis/app/helpers/ios/MongoIOTest.java
@@ -1,0 +1,20 @@
+package net.yuwtennis.app.helpers.ios;
+
+import org.junit.Test;
+import static org.junit.Assert.assertTrue;
+
+public class MongoIOTest {
+
+    @Test
+    public void Write() {
+        // TODO FIXME
+        assertTrue(true) ;
+    }
+
+    @Test
+    public void Read() {
+        // TODO FIXME
+        assertTrue(true) ;
+    }
+
+}
